fixed hex printing

+++ Raspi/main.cpp
@@ -120,7 +120,7 @@ int main(int argc, char** argv) {
         if (key.codepoint == 0 && key.key == 0)
             continue;
         std::cout << "Key U+" << std::setfill('0') << std::setw(6) << std::right << std::hex << key.codepoint
-                  << " ASCII 0x" << std::setfill('0') << std::setw(2) << std::right << std::hex << key.key << std::endl;
+                  << " ASCII 0x" << std::setfill('0') << std::setw(2) << std::right << std::hex << static_cast<int>(key.key) << std::endl;
     }
 
     // while (true) {
